Update States to Bevy 0.10.
