+++ src/code/src/basics.rs
@@ -1361,162 +1361,6 @@ fn main() {
 }
 
 #[allow(dead_code)]
-#[allow(unused_imports)]
-mod app5 {
-    use bevy::prelude::*;
-    use super::*;
-
-// ANCHOR: check-state
-fn play_music(
-    app_state: Res<State<AppState>>,
-    // ...
-) {
-    match app_state.current() {
-        AppState::MainMenu => {
-            // TODO: play menu music
-        }
-        AppState::InGame => {
-            // TODO: play game music
-        }
-        AppState::Paused => {
-            // TODO: play pause screen music
-        }
-    }
-}
-// ANCHOR_END: check-state
-
-// ANCHOR: change-state
-fn enter_game(mut app_state: ResMut<State<AppState>>) {
-    app_state.set(AppState::InGame).unwrap();
-    // ^ this can fail if we are already in the target state
-    // or if another state change is already queued
-}
-// ANCHOR_END: change-state
-
-fn pushpop(mut app_state: ResMut<State<AppState>>) {
-// ANCHOR: state-push-pop
-    // to go into the pause screen
-    app_state.push(AppState::Paused).unwrap();
-    // to go back into the game
-    app_state.pop().unwrap();
-// ANCHOR_END: state-push-pop
-}
-
-// ANCHOR: state-input-clear
-fn esc_to_menu(
-    mut keys: ResMut<Input<KeyCode>>,
-    mut app_state: ResMut<State<AppState>>,
-) {
-    if keys.just_pressed(KeyCode::Escape) {
-        app_state.set(AppState::MainMenu).unwrap();
-        keys.reset(KeyCode::Escape);
-    }
-}
-// ANCHOR_END: state-input-clear
-
-// ANCHOR: app-states
-#[derive(Debug, Clone, Eq, PartialEq, Hash)]
-enum AppState {
-    MainMenu,
-    InGame,
-    Paused,
-}
-
-fn main() {
-    App::new()
-        .add_plugins(DefaultPlugins)
-
-        // add the app state type
-        .add_state(AppState::MainMenu)
-
-        // add systems to run regardless of state, as usual
-        .add_system(play_music)
-
-        // systems to run only in the main menu
-        .add_system_set(
-            SystemSet::on_update(AppState::MainMenu)
-                .with_system(handle_ui_buttons)
-        )
-
-        // setup when entering the state
-        .add_system_set(
-            SystemSet::on_enter(AppState::MainMenu)
-                .with_system(setup_menu)
-        )
-
-        // cleanup when exiting the state
-        .add_system_set(
-            SystemSet::on_exit(AppState::MainMenu)
-                .with_system(close_menu)
-        )
-        .run();
-}
-// ANCHOR_END: app-states
-
-    fn animate_trees() {}
-    fn animate_water() {}
-    fn player_movement() {}
-    fn player_idle() {}
-    fn reset_player() {}
-    fn hide_enemies() {}
-    fn setup_player() {}
-    fn despawn_player() {}
-    fn setup_map() {}
-    fn despawn_map() {}
-    fn handle_ui_buttons() {}
-    fn setup_menu() {}
-    fn close_menu() {}
-
-fn main2() {
-    App::new()
-        .add_plugins(DefaultPlugins)
-        // add the app state type
-        .add_state(AppState::InGame)
-// ANCHOR: state-stack
-        // player movement only when actively playing
-        .add_system_set(
-            SystemSet::on_update(AppState::InGame)
-                .with_system(player_movement)
-        )
-        // player idle animation while paused
-        .add_system_set(
-            SystemSet::on_inactive_update(AppState::InGame)
-                .with_system(player_idle)
-        )
-        // animations both while paused and while active
-        .add_system_set(
-            SystemSet::on_in_stack_update(AppState::InGame)
-                .with_system(animate_trees)
-                .with_system(animate_water)
-        )
-        // things to do when becoming inactive
-        .add_system_set(
-            SystemSet::on_pause(AppState::InGame)
-                .with_system(hide_enemies)
-        )
-        // things to do when becoming active again
-        .add_system_set(
-            SystemSet::on_resume(AppState::InGame)
-                .with_system(reset_player)
-        )
-        // setup when first entering the game
-        .add_system_set(
-            SystemSet::on_enter(AppState::InGame)
-                .with_system(setup_player)
-                .with_system(setup_map)
-        )
-        // cleanup when finally exiting the game
-        .add_system_set(
-            SystemSet::on_exit(AppState::InGame)
-                .with_system(despawn_player)
-                .with_system(despawn_map)
-        )
-// ANCHOR_END: state-stack
-        .run();
-}
-}
-
-#[allow(dead_code)]
 mod app6 {
 use bevy::prelude::*;
